compression.rs: define arithmetic operations on Compressed2
Also Clone, PartialEq, Eq.

Change performance test to take a compressed version as argument,
compress once before the thread test (decompress to verify to be sure
it's right before arithmetic), then modify the numbers in the loops
(this was to verify that caches don't omit writing to memory when
contents don't change).